requirements.txt
msal
requests
beautifulsoup4
tqdm
python-dotenv
